Fixed Update Comment

+++ backend/services/src/main/java/com/crimealert/controllers/CommentController.java
@@ -7,6 +7,8 @@ import org.bson.Document;
 import com.crimealert.Exceptions.ClientSideException;
 import com.crimealert.models.Comment;
 import com.crimealert.services.CommentService;
+import com.crimealert.utils.ValidatorUtil;
+
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
@@ -52,14 +54,18 @@ public class CommentController {
 			 Comment comment
 	) {
     	try {
-    	
+
+    		if ((comment.getComment() != null && !comment.getComment().isEmpty()) && (comment.getId() != null && !comment.getId().isEmpty())) {
     			
         	Document response = getCommentService().updateComment(comment);
         		
         	System.out.println("Post Update Response : " + response.toJson());
-        		
+        	
         	return Response.ok(response.toJson()).build();
-    		
+    		} else {
+    			return Response.status(400).entity("Comment or Comment Id cannot be empty").build();
+    		}
+        		
     	} catch (ClientSideException ex) {
     		return Response.status(400).entity(ex.getMessage()).build();
     	} catch (Exception ex) {
@@ -68,19 +74,24 @@ public class CommentController {
 	}
 	
 	@DELETE
-    @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
 	public Response deleteComment(
 			@PathParam("commentId")String commentId
 	) {
     	try {
     	
+    		if (commentId != null && !commentId.isEmpty()){
     			
         	Document response = getCommentService().deleteComment(commentId);
         		
         	System.out.println("Post Deletion Response : " + response.toJson());
         		
         	return Response.ok(response.toJson()).build();
+    		}
+    		else {
+    			return Response.status(400).entity("Comment Id cannot be empty").build();
+    		}
+    		
     		
     	} catch (ClientSideException ex) {
     		return Response.status(400).entity(ex.getMessage()).build();
